Изучение модуля re для обработки регулярный выражений

+++ logical_operators.py
@@ -1,4 +1,4 @@
-number = int(input("Enter a number: "))
+number = float(input("Enter a number: "))
 if number > 5:
     print("This number is greater than five.")
